# docs/swagger.yaml
basePath: /
definitions:
  controllers.ExpenseResponse:
    properties:
      expense:
        type: string
    type: object
  controllers.IncomeResponse:
    properties:
      income:
        type: string
    type: object
...
  controllers.TransactionHistoryResponse:
    properties:
      amount:
        description: Jumlah transaksi
        type: number
      created_at:
        description: Timestamp transaksi
        type: string
      id:
        description: ID transaksi
        type: integer
      related_user_fullname:
        description: Nama lengkap pengguna yang terkait
        type: string
      related_user_image:
        description: Gambar pengguna yang terkait
        type: string
      related_user_phone:
        description: Nomor telepon pengguna yang terkait
        type: string
      transaction_type:
        description: '''Sent'', ''Received'', ''Top-Up'''
        type: string
      user_fullname:
        description: Nama lengkap pengguna
        type: string
      user_image:
        description: Gambar pengguna
        type: string
      user_phone:
        description: Nomor telepon pengguna
        type: string
    type: object
  dto.LoginDTO:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  dto.PinDTO:
    properties:
      pin:
        type: string
    type: object
  dto.RegisterDTO:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
...
  models.PaymentMethod:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      isDeleted:
        type: boolean
      name:
        type: string
      tax:
        type: number
      topupTransactions:
        items:
          $ref: '#/definitions/models.TopupTransaction'
        type: array
      updatedAt:
        type: string
    type: object
  models.TopupTransaction:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      isDeleted:
        type: boolean
      payment_method_id:
        type: integer
      transactionID:
        type: integer
      updatedAt:
        type: string
    type: object
info:
  contact: {}
  description: This is example for server P-Pay Aplication
  title: Backend P-Pay
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Login Account
      parameters:
      - description: Input Email
        in: formData
        name: email
        required: true
        type: string
      - description: Input Password
        in: formData
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.LoginDTO'
      tags:
      - Auth
  /auth/pin:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Authentication Pin
      parameters:
      - description: Input Pin
        in: formData
        name: pin
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.PinDTO'
      security:
      - ApiKeyAuth: []
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Registrasi Account
      parameters:
      - description: Input Email
        in: formData
        name: email
        required: true
        type: string
      - description: Input Password
        in: formData
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.RegisterDTO'
      tags:
      - Auth
...
  /transaction/expense:
    get:
      description: get user expense
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.ExpenseResponse'
      security:
      - ApiKeyAuth: []
      summary: get user expense
      tags:
      - Transaction
  /transaction/history:
    get:
      description: Get Transaction History
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.TransactionHistoryResponse'
      security:
      - ApiKeyAuth: []
      summary: get history transaction
      tags:
      - Transaction
  /transaction/income:
    get:
      description: get user income
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.IncomeResponse'
      security:
      - ApiKeyAuth: []
      summary: get user income
      tags:
      - Transaction
  /transaction/payment:
    get:
      description: get payment method
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PaymentMethod'
      security:
      - ApiKeyAuth: []
      summary: get payment method
      tags:
      - Transaction
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"